+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Refactored citation section to use a record instance as a single prop
 * Refactored resource type to use an enum
 * Refactoring ideas into a data file and adding cancelled state
+* Refactoring build information out of epilogue component into utils module
 
 ## [0.5.3] - 2023-12-30